Toggle for production and staging HACI in showcase app (#149)
This allows enable and disable dev mode when using the Showcase app. The only feature impacted at this moment is the HACI, but the implementation allows extending this for other features in the future. The dev mode configuration is not persistent, so when you (re)launch the app, it will always be set to production.